Separation of the UserController from the UsersController. UserController POST can return Conflict
